requirements.txt
clip @ git+https://github.com/openai/CLIP.git
decord==0.6.0
einops==0.7.0
finch-clust==0.2.0
hydra-core==1.3.2
hyperlight==0.0.5
imageio==2.23.0
jupyter
kornia==0.7.0
lightning==2.1.0
matplotlib
numpy==1.19.2
omegaconf
opencv-python
pandas
panopticapi @ git+https://github.com/cocodataset/panopticapi.git
pillow
pycocotools
pytorch-lightning==2.1.0
pytorch-msssim==0.2.1
scikit-learn==1.3.2
scipy==1.11.3
tensorboard==2.14.0
timm==0.6.7
torch==2.0.1
torchmetrics==1.2.0
torchvision==0.15.2